docs: finalização e comentarios do projeto

+++ server.py
@@ -1,7 +1,6 @@
 import os
 import json
 from http.server import SimpleHTTPRequestHandler, HTTPServer
-from urllib.parse import parse_qs
 
 """A definição do handler personalizado é criado através de uma classe que herda o "SimpleHTTPRequestHandler".
 O objetivo receber e processar as respostas de um evento específico que ocorre dentro do servidor."""
@@ -11,6 +10,7 @@ json_usuarios_cadastrados = "./data/users.json"
 
 class MyHandle(SimpleHTTPRequestHandler):
     def list_directory(self, path):
+        """ Função utilizada para renderizar a primeira página """
         try:
             f = open(os.path.join(path, "./templates/index.html"), encoding="utf-8")
             self.send_response(200)
@@ -20,11 +20,12 @@ class MyHandle(SimpleHTTPRequestHandler):
             f.close()
             return None
         except FileNotFoundError:
-            pass 
+            self.send_error(404, "File Not Found")
         return super().list_directory(path)
 
     def carregar_pagina(self, caminho):
         try:
+            """ Função simples para reutilização de código, utilizada nos metódos GET para carregar a página """
             with open(os.path.join(os.getcwd(), caminho), "r", encoding="utf-8") as arquivo:
                 content = arquivo.read()
                 self.send_response(200)
@@ -33,13 +34,6 @@ class MyHandle(SimpleHTTPRequestHandler):
                 self.wfile.write(content.encode("utf-8"))
         except FileNotFoundError:
             self.send_error(404, f"Arquivo não {caminho} encontrado")
-
-    def retornar_post_formulario(self):
-        """ Função usada para ler e retornar o corpo do POST já decodificado."""
-        content_length = int(self.headers["Content-length"])
-        body = self.rfile.read(content_length).decode("utf-8")
-        form_data = parse_qs(body)
-        return form_data
 
     def do_GET(self):
         try:    
@@ -74,13 +68,14 @@ class MyHandle(SimpleHTTPRequestHandler):
                     return {FileNotFoundError: "Caminho não encontrado!"}
             else:
                 super().do_GET()
+
         except FileNotFoundError:
             self.send_error(404, "File Not Found")
     
     def do_POST(self):
-        """ Os dois metódos POST abaixo funcionam da mesma forma, recebendo informações dos formulário
-        e realizando questões como autenticação e liberação. """
+        """ Os dois metódos POST abaixo funcionam da mesma forma, recebendo informações dos formulário do java script e realizando questões como autenticação e liberação. """
         if self.path == "/send_login":
+            """ Recebe a requisição do fetch do java script e armazena as informações na variável 'data' """
             content_lenght = int(self.headers.get("Content-Length", 0))
             body = self.rfile.read(content_lenght).decode("utf-8")
             data = json.loads(body)
@@ -109,7 +104,7 @@ class MyHandle(SimpleHTTPRequestHandler):
                     self.end_headers()
                     self.wfile.write(json.dumps({"message": "usuário ou senha inválidos"}).encode("utf-8"))
             else:
-                print({"error": "problema ao encontrar o arquivo de usuários"})
+                self.send_error(404, "File not found")
 
         elif self.path == "/send_cadastro":
             content_lenght = int(self.headers.get("Content-Length", 0))
@@ -123,67 +118,75 @@ class MyHandle(SimpleHTTPRequestHandler):
             valid_user = False
 
             if os.path.exists(json_usuarios_cadastrados):
-                    if password_form == confirm_password_form:
-                        with open(json_usuarios_cadastrados, "r", encoding="utf-8") as arquivo_json:
-                            users = json.load(arquivo_json)
+                if password_form == confirm_password_form:
+                    with open(json_usuarios_cadastrados, "r", encoding="utf-8") as arquivo_json:
+                        users = json.load(arquivo_json)
 
-                        id_user = len(users) + 1
-                        user = {"id": id_user, "user": user_form, "password": password_form}
-                        users.append(user)
-                        
-                        with open(json_usuarios_cadastrados, "w", encoding="utf-8") as arquivo_json:
-                            json.dump(users, arquivo_json, indent=4, ensure_ascii=False)
-                                
-                        valid_user = True
+                    """ Pequena lógica para permitir a inserção do novo usuário no json """
+                    id_user = len(users) + 1
+                    user = {"id": id_user, "user": user_form, "password": password_form}
+                    users.append(user)
+                    
+                    with open(json_usuarios_cadastrados, "w", encoding="utf-8") as arquivo_json:
+                        json.dump(users, arquivo_json, indent=4, ensure_ascii=False)
+                            
+                    valid_user = True
             
-            if valid_user:
-                self.send_response(200)
-                self.send_header("Content-Type", "application/json")
-                self.end_headers()
-                self.wfile.write(json.dumps({"message": "ok"}).encode("utf-8"))
-            else:
-                self.send_response(400)
-                self.send_header("Content-Type", "application/json")
-                self.end_headers()
-                self.wfile.write(json.dumps({"message": "credenciais incorretas para realizar o cadastro"}).encode("utf-8"))
+                    if valid_user:
+                        self.send_response(200)
+                        self.send_header("Content-Type", "application/json")
+                        self.end_headers()
+                        self.wfile.write(json.dumps({"message": "ok"}).encode("utf-8"))
+                    else:
+                        self.send_response(400)
+                        self.send_header("Content-Type", "application/json")
+                        self.end_headers()
+                        self.wfile.write(json.dumps({"message": "credenciais incorretas para realizar o cadastro"}).encode("utf-8"))
+                else:
+                    self.send_error(404, "File not found")
 
         elif self.path == "/send_movie":
             content_lenght = int(self.headers.get("Content-Length", 0))
             body = self.rfile.read(content_lenght).decode("utf-8")
             data = json.loads(body)
             
-            with open(json_filmes_cadastrados, "r", encoding="utf-8") as arquivo_json:
-                movie = json.load(arquivo_json)
+            if os.path.exists(json_filmes_cadastrados):
+                with open(json_filmes_cadastrados, "r", encoding="utf-8") as arquivo_json:
+                    movie = json.load(arquivo_json)
 
-            id_movie = len(movie) + 1
-            data = {"id": id_movie, **data}
-            movie.append(data)
+                """ Pequena lógica para permitir a inserção de um novo filme no json """
+                id_movie = len(movie) + 1
+                data = {"id": id_movie, **data}
+                movie.append(data)
 
-            with open(json_filmes_cadastrados, "w", encoding="utf-8") as arquivo_json:
-                json.dump(movie, arquivo_json, indent=4, ensure_ascii=False)
+                with open(json_filmes_cadastrados, "w", encoding="utf-8") as arquivo_json:
+                    json.dump(movie, arquivo_json, indent=4, ensure_ascii=False)
 
-            self.send_response(200)
-            self.send_header("Content-Type", "application/json")
-            self.end_headers()
-            self.wfile.write(json.dumps({"message": "ok"}).encode("utf-8"))
-    
+                self.send_response(200)
+                self.send_header("Content-Type", "application/json")
+                self.end_headers()
+                self.wfile.write(json.dumps({"message": "ok"}).encode("utf-8"))
+            else:
+                self.send_error(404, "File not found")
+        else:
+            super().do_GET()
+
     def do_PUT(self):
-        path = self.path.split("/")
-        if f"/{path[1]}" == "/update_movie":
+        path = self.path.split("/") # Busca os path parameters da requisição e os divide em uma lista, sendo o caminho em si e o id do filme
+        if f"/{path[1]}" == "/update_movie": # Verfica se o caminho corresponde ao de atualização de filmes
             content_lenght = int(self.headers.get("Content-Length", 0))
             body = self.rfile.read(content_lenght).decode("utf-8")
             data = json.loads(body)
 
-            movie_id = path[2]
+            movie_id = path[2] # Pega o id do filme de acordo com a lista gerada pelo split do caminho
             try:
                 with open(json_filmes_cadastrados, "r", encoding="utf-8") as arquivo_json:
                     filmes = json.load(arquivo_json)
 
-
+                """ For para verificar qual filme corresponde ao filme editado """
                 for filme in filmes:
-                    if filme['id'] == movie_id:
+                    if filme['id'] == int(movie_id):
                         filme.update(data)
-                
 
                 with open(json_filmes_cadastrados, "w", encoding="utf-8") as arquivo_json:
                     json.dump(filmes, arquivo_json, indent=4, ensure_ascii=False)
@@ -197,34 +200,35 @@ class MyHandle(SimpleHTTPRequestHandler):
                 self.send_response(500)
                 self.end_headers()
                 self.wfile.write(json.dumps({"message": "erro ao ler o arquivo JSON."}).encode("utf-8"))
-            
+        else:
+            super().do_GET()
+
     def do_DELETE(self): 
         url = self.path
         id_movie = url.split('/')[2]
 
-        try:
-            with open(json_filmes_cadastrados, "r", encoding="utf-8") as arquivo_json:
-                filmes = json.load(arquivo_json)
-            
-            filmes_novos = [filme for filme in filmes if filme['id'] != id_movie]
+        if url.split('/')[1] == "delete_movie":
+            try:
+                with open(json_filmes_cadastrados, "r", encoding="utf-8") as arquivo_json:
+                    filmes = json.load(arquivo_json)
+                
+                filmes_novos = [filme for filme in filmes if filme['id'] != int(id_movie)]
 
-            with open(json_filmes_cadastrados, "w", encoding="utf-8") as arquivo_json:
-                json.dump(filmes_novos, arquivo_json, indent=4, ensure_ascii=False)
+                with open(json_filmes_cadastrados, "w", encoding="utf-8") as arquivo_json:
+                    json.dump(filmes_novos, arquivo_json, indent=4, ensure_ascii=False)
 
-            if len(filmes) > len(filmes_novos):
-                self.send_response(200)
-                self.send_header("Content-Type", "application/json")
+                if len(filmes) > len(filmes_novos):
+                    self.send_response(200)
+                    self.send_header("Content-Type", "application/json")
+                    self.end_headers()
+                    self.wfile.write(json.dumps({"message": "filme deletado com sucesso"}).encode("utf-8"))
+            except (json.JSONDecodeError):
+                self.send_response(500)
                 self.end_headers()
-                self.wfile.write(json.dumps({"message": "filme deletado com sucesso"}).encode("utf-8"))
-        except (json.JSONDecodeError):
-            self.send_response(500)
-            self.end_headers()
-            self.wfile.write(json.dumps({"message": "erro ao ler o arquivo JSON."}).encode("utf-8"))
+                self.wfile.write(json.dumps({"message": "erro ao ler o arquivo JSON."}).encode("utf-8"))
             
 def main():
-    """Função para iniciar o servidor, recebe a porta que deve ser utilizada, ou seja , o endereço do servidor, e o handle personalidado criado na classe
-    acima.
-    """
+    """Função para iniciar o servidor, recebe a porta que deve ser utilizada, ou seja , o endereço do servidor, e o handle personalidado criado na classe acima. """
     server_address = ("", 8000)
     httpd = HTTPServer(server_address, MyHandle)
     print(f"Servidor rodando na porta http://localhost:{server_address[1]}") # Os colchetes no server_address é utilizado para pegar a porta indicada no código
